Adjust route and move folders
