+++ app/backend/src/couchers/urls.py
@@ -39,7 +39,7 @@ def account_settings_link():
     return f"{config['BASE_URL']}/account-settings"
 
 def notification_settings_link():
-    return f"{config['BASE_URL']}/notification-settings"
+    return f"{config['BASE_URL']}/account-settings/notification-settings"
 
 def feature_preview_link():
     return f"{config['BASE_URL']}/preview"
