+++ app/web/components/CustomColorSwitch.tsx
