+++ app/web/pages/account-settings/notifications.tsx
@@ -1,5 +1,5 @@
 import { appGetLayout } from "components/AppRoute";
-import EditNotificationSettingsPageComponent from "features/profile/edit/notificationSettings/EditNotificationSettingsPage";
+import EditNotificationSettingsPageComponent from "features/auth/notifications/EditNotificationSettingsPage";
 import { GetStaticProps } from "next";
 import nextI18nextConfig from "next-i18next.config";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
